requirement.txt
langchain
langchain-core
langchain-community
langchain-mistralai
python-dotenv
chromadb
pypdf
beautifulsoup4
langchain-huggingface
sentence-transformers
